% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hr.R
\name{calc_hr}
\alias{calc_hr}
\title{Generate hazard ratio with prediction intervals from trial simulation}
\usage{
calc_hr(sim)
}
\description{
Generate hazard ratio with prediction intervals from trial simulation
}
