% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_km.R
\name{calc_km}
\alias{calc_km}
\title{Generate statistics from simulation runs on Kaplan-Meier medians}
\usage{
calc_km(sim, ci.range = 0.95)
}
\arguments{
\item{sim}{A \code{trialsim} class object generated by \code{\link[=trial_sim]{trial_sim()}} function.}

\item{ci.range}{Prediction interval for simulated survival KM medians (defaults to 0.95).}
}
\value{
A \code{trialsim.km} object that contains a data frame (tibble) for the KM medians of each simulation with the
following columns:
\itemize{\item \strong{rep}: ID for simulation runs
\item \strong{rx}: Treatment group
\item \strong{median}: KM median for each simulation
}
And a data frame containing the summary statistics for KM medians across simulations:
\itemize{\item \strong{rx}: Treatment group
\item \strong{description}: Labels for KM statistics: lower limit \ median \ upper limit
\item \strong{KM_median}: Statistic for KM medians
\item \strong{quantile}: Quantile of simulations (corresponding to \code{ci.range} parameter)
}
}
\description{
Generate statistics from simulation runs on Kaplan-Meier medians
}
\details{
Extracts median and prediction interval by treatment group across all repeated simulations for median survival.
}
\examples{
sim.km <- calc_km(sim)
}
