% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_sim_plot.R
\name{plot_sim}
\alias{plot_sim}
\alias{plot_sim.trialsim.km}
\alias{plot_sim.trialsim.hr}
\title{Generic graph function}
\usage{
plot_sim(sim, ci, summary)

\method{plot_sim}{trialsim.km}(sim, ci = FALSE, summary = FALSE)

\method{plot_sim}{trialsim.hr}(sim, ci = FALSE, summary = FALSE)
}
\arguments{
\item{sim}{An object from \code{\link[=calc_km]{calc_km()}} or \code{\link[=calc_hr]{calc_hr()}}}

\item{ci}{A logical specifying whether confidence bounds are shown on the histogram plot (default=FALSE)}

\item{summary}{A logical specifying whether a summary table is presented on the graph (default=FALSE)}
}
\description{
Plots simulation results for either Kaplan-Meier medians by treatment group, or hazard ratios generated
from simulated trials.
}
\section{Methods (by class)}{
\itemize{
\item \code{trialsim.km}: Simulation results presented for Kaplan-Meier medians

\item \code{trialsim.hr}: Simulation results presented for hazard ratios
}}

