% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_sim.R
\name{trial_sim}
\alias{trial_sim}
\title{Simulation of trial data for Sip-T project}
\usage{
trial_sim(schedule, enrol, rxrate, censor.prop = 0.1, n.rep = 1000)
}
\arguments{
\item{schedule}{A required vector describing the visit schedule in weeks.}

\item{enrol}{A required vector that displays the number of subjects projected to be enrolled each
month.  It incorporates information on the ramp-up period (see example below).}

\item{rxrate}{A required vector providing the median PFS for each of two groups
(reference / treatment).}

\item{censor.prop}{Proportion of subjects censored in simulation (e.g. 0.1 for 10\%)}

\item{n.rep}{An integer defining number of simulations}
}
\value{
A \code{survparamsim} object that contains a data frame (tibble) for the simulated survival profiles with the
following columns:
\itemize{\item \strong{rep}: ID for simulation runs
\item \strong{subjid}: ID for subjects in simulations
\item \strong{pfst}: PFS time
\item \strong{status}: event status, 0=censored, 1=event
}
}
\description{
Simulation of trial data for Sip-T project
}
