% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_km.R
\name{calc_km}
\alias{calc_km}
\title{Generate Kaplan-Meier curves with prediction intervals from simulation routine}
\usage{
calc_km(sim)
}
\arguments{
\item{sim}{A \code{trialsim} class object generated by \code{\link[=trial_sim]{trial_sim()}} function.}
}
\description{
Generate Kaplan-Meier curves with prediction intervals from simulation routine
}
